Fix a most grievous error
In response to the following email:

Salutations,

In the recent article you seem to have overlooked
a grievous error while adding links to the page.
In the content section, you have listed Spinning Galaxy
no less than 2 times.  Also, just before that,
you show 6 pictures.  It is very suspicious that
the first five pictures correspond to the first five
links listed.  One might draw the conclusion that
you intended to provide a link to a sixth, maze like
example.  I hope you can imagine my horror when
discover that the sixth link provide was not
the advertised example, but seems to be a duplicate
of the Spinning Galaxy example mentioned
previously.

So please pardon me, should this be mere malice on
your part and not a simple error, however I must
implore and beseech you to correct it.

Best Regards,
Leland